html/browsers/browsing-the-web/history-traversal/Event-hashchange.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>history-traversal Test: event - hashchange</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<meta name="assert" content="Check if event hashchange is fired When the hash is changed.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>

async_test(function (t) {
  var oldURL = window.location.href;

  window.onhashchange = t.step_func(function (e) {
    assert_true(e.isTrusted);
    assert_equals(e.target, window);
    assert_equals(e.type, "hashchange");
    assert_true(e instanceof HashChangeEvent);
    assert_equals(e.oldURL, oldURL);
    assert_equals(e.newURL, window.location.href);
    assert_true(e.bubbles);
    assert_false(e.cancelable);
    t.done();
  });
  window.location.hash = "#newHash";
});

</script>
